# [Profiler] Pay for what you use (v2)