--- a/torch/csrc/profiler/collection.h
+++ b/torch/csrc/profiler/collection.h
@@ -6,6 +6,7 @@
 
 #include <c10/macros/Macros.h>
 #include <c10/util/flat_hash_map.h>
+#include <c10/util/variant.h>
 #include <torch/csrc/profiler/containers.h>
 #include <torch/csrc/profiler/kineto_shim.h>
 #include <torch/csrc/profiler/util.h>
@@ -14,18 +15,9 @@ namespace torch {
 namespace profiler {
 namespace impl {
 
-// `reportBackendEventToActiveKinetoProfiler` reports times rather than counts,
-// so `OpEventData` has to be able to store both cases.
-union TimeStamp {
-  int64_t us_; // Backend event.
-  approx_time_t count_;
-};
-
-// NOLINTNEXTLINE(cppcoreguidelines-pro-type-member-init)
-struct TORCH_API OpEventData {
-  OpEventData() = default;
-
-  OpEventData(
+struct OpEvent {
+  OpEvent() = default;
+  OpEvent(
       const uint64_t correlation_id,
       const uint64_t start_thread_id,
       const int64_t sequence_number,
@@ -41,36 +33,10 @@ struct TORCH_API OpEventData {
         record_function_scope_{(uint8_t)scope},
         is_async_{is_async},
         debug_handle_{debug_handle},
-        kineto_info_{kineto::kineto_ids()},
-        name_{std::move(name)} {
-    end_time_.count_ = std::numeric_limits<approx_time_t>::min();
-  }
+        name_{name} {}
 
-  OpEventData(
-      const int64_t start_time,
-      const int64_t end_time,
-      const at::RecordScope scope,
-      const int64_t debug_handle,
-      const std::string name,
-      const std::string backend)
-      : correlation_id_{std::numeric_limits<uint64_t>::max()},
-        start_thread_id_{at::RecordFunction::currentThreadId()},
-        end_thread_id_{start_thread_id_},
-        sequence_number_{-1},
-        forward_thread_id_{start_thread_id_},
-        record_function_scope_{(uint8_t)scope},
-        is_async_{false},
-        debug_handle_{debug_handle},
-        kineto_info_{kineto::kineto_ids()},
-        name_{std::move(name)},
-        backend_{std::move(backend)} {
-    start_time_.us_ = start_time;
-    end_time_.us_ = end_time;
-  }
-
-  // POD members
-  TimeStamp start_time_;
-  TimeStamp end_time_;
+  approx_time_t start_time_;
+  approx_time_t end_time_{std::numeric_limits<approx_time_t>::min()};
   uint64_t correlation_id_;
   uint64_t start_thread_id_;
   uint64_t end_thread_id_;
@@ -79,54 +45,116 @@ struct TORCH_API OpEventData {
   uint8_t record_function_scope_;
   bool is_async_;
   int64_t debug_handle_;
-  kineto::DeviceAndResource kineto_info_;
-
   std::string name_;
+};
 
-  // report_input_shapes
+struct Inputs {
   std::vector<std::vector<int64_t>> shapes_;
   std::vector<std::string> dtypes_;
+};
 
-  // with_stack
-  std::vector<std::string> stack_;
+struct FallbackPair {
+  CUDAEventStub cuda_event_start_ = nullptr;
+  CUDAEventStub cuda_event_end_ = nullptr;
+};
 
-  // with_modules
-  c10::optional<std::vector<std::string>> module_hierarchy_;
+struct BackendEvent {
+  int64_t start_time_us_;
+  int64_t end_time_us_;
+  uint8_t record_function_scope_;
+  int64_t debug_handle_;
+  std::string name_;
+  std::string backend_;
+};
 
-  // with_flops
+// NOLINTNEXTLINE(cppcoreguidelines-pro-type-member-init)
+struct Result {
+  std::string name() const;
+  uint64_t correlation_id() const;
+
+  int64_t start_time_us_;
+  int64_t end_time_us_;
+  uint64_t start_tid_;
+  kineto::DeviceAndResource kineto_info_;
+
+  c10::variant<OpEvent, BackendEvent> event_;
+
+  // OpEvent only.
+  Inputs inputs_;
+  std::vector<std::string> jit_stack_;
+  std::vector<std::string> jit_modules_;
   std::unordered_map<std::string, c10::IValue> extra_args_;
+  FallbackPair gpu_fallback_;
+};
 
-  // reportBackendEventToActiveKinetoProfiler
-  c10::optional<std::string> backend_;
+struct KinetoObserverContext : public at::ObserverContext {
+  explicit KinetoObserverContext(OpEvent* event)
+    : event_{event} {}
 
-  // ProfilerState::KINETO_GPU_FALLBACK
-  torch::profiler::impl::CUDAEventStub cuda_event_start_ = nullptr;
-  torch::profiler::impl::CUDAEventStub cuda_event_end_ = nullptr;
+  OpEvent* event_;
+  FallbackPair* fallback_ {nullptr};
 };
 
 class TORCH_API ThreadLocalSubqueue {
  public:
+  ThreadLocalSubqueue(const uint64_t tid, const ProfilerConfig& config);
+
+  std::unique_ptr<KinetoObserverContext> begin_op(const at::RecordFunction& fn, uint64_t correlation_id);
+
   template <class... Args>
-  OpEventData* emplace_back(Args&&... args) {
-    return data_.emplace_back(std::forward<Args>(args)...);
+  void emplace_backend_event(Args&&... args) {
+    backend_events_.emplace_back(std::forward<Args>(args)...);
+  }
+
+  uint64_t tid() const {
+    return tid_;
+  }
+
+  const kineto::DeviceAndResource& kineto_info() const {
+    return kineto_info_;
   }
 
  private:
-  friend class RecordQueue;
+  uint64_t tid_;
+  ProfilerConfig config_;
+  kineto::DeviceAndResource kineto_info_;
 
+  friend class RecordQueue;
   // See `containers.h` for block size benchmarks.
-  static constexpr size_t BlockSize = 1024;
-  AppendOnlyList<OpEventData, BlockSize> data_;
+  static constexpr size_t BlockSize = 512;
+  AppendOnlyList<OpEvent, BlockSize> op_events_;
+
+  // report_input_shapes
+  AppendOnlyList<Inputs, BlockSize> inputs_;
+
+  // with_stack
+  AppendOnlyList<std::vector<std::string>, BlockSize> jit_stack_;
+
+  // with_modules
+  AppendOnlyList<std::vector<std::string>, BlockSize> jit_modules_;
+
+  // with_flops
+  AppendOnlyList<std::unordered_map<std::string, c10::IValue>, BlockSize> extra_args_;
+
+  // ProfilerState::KINETO_GPU_FALLBACK
+  AppendOnlyList<FallbackPair, BlockSize> gpu_fallback_;
+
+  // reportBackendEventToActiveKinetoProfiler
+  AppendOnlyList<BackendEvent, BlockSize> backend_events_;
 };
 
 class TORCH_API RecordQueue {
  public:
-  RecordQueue();
+  explicit RecordQueue(const ProfilerConfig& config);
+
   ThreadLocalSubqueue* getSubqueue();
-  std::deque<OpEventData> getRecords(std::function<time_t(approx_time_t)> time_converter);
+
+  // NB: This is a destructive operation.
+  std::deque<Result> getRecords(std::function<time_t(approx_time_t)> time_converter);
 
  private:
   uint32_t id_;
+  ProfilerConfig config_;
   ska::flat_hash_map<uint64_t, std::unique_ptr<ThreadLocalSubqueue>> sub_queues_;
   std::mutex sub_queue_mutex_;
 };